feat: merge stake now
