--- packages/web/src/components/organisms/StakeNow/index.tsx
@@ -1,20 +1,0 @@
-import * as React from 'react'
-import { useStake } from 'src/fixtures/dev-kit/hooks'
-import { InputFormCard } from 'src/components/molecules/InputFormCard'
-
-interface Props {
-  propertyAddress: string
-}
-
-export const StakeNow = ({ propertyAddress }: Props) => {
-  const { stake } = useStake()
-
-  const handleSubmit = React.useCallback(
-    (amount: string) => {
-      stake(propertyAddress, amount)
-    },
-    [stake, propertyAddress]
-  )
-
-  return <InputFormCard label="Stake Now" suffix="DEV" onSubmitStake={handleSubmit} />
-}
