+++ packages/web/src/components/organisms/TransactionForm/index.tsx
@@ -2,6 +2,7 @@ import React, { useCallback } from 'react'
 import { Space } from 'antd'
 import { useGetMyStakingAmount, useWithdrawStakingReward } from 'src/fixtures/dev-kit/hooks'
 import { useWithdrawHolderReward, useGetMyHolderAmount } from 'src/fixtures/dev-kit/hooks'
+import { useStake } from 'src/fixtures/dev-kit/hooks'
 import { WithdrawCard } from 'src/components/molecules/WithdrawCard'
 import { InputFormCard } from 'src/components/molecules/InputFormCard'
 import { useGetLastAllocatorAllocationResultQuery } from '@dev/graphql'
@@ -11,16 +12,24 @@ interface Props {
 }
 
 export const TransactionForm = ({ propertyAddress }: Props) => {
+  const { stake } = useStake()
   const { withdrawStaking } = useWithdrawStakingReward()
   const { withdrawHolder } = useWithdrawHolderReward()
   const { myStakingAmount } = useGetMyStakingAmount(propertyAddress)
   const { myHolderAmount } = useGetMyHolderAmount(propertyAddress)
   const { data } = useGetLastAllocatorAllocationResultQuery({ variables: { propertyAddress } })
+  const handleSubmit = React.useCallback(
+    (amount: string) => {
+      stake(propertyAddress, amount)
+    },
+    [stake, propertyAddress]
+  )
   const handleWithdrawStaking = useCallback(() => withdrawStaking(propertyAddress), [propertyAddress, withdrawStaking])
   const handleWithdrawHolder = useCallback(() => withdrawHolder(propertyAddress), [propertyAddress, withdrawHolder])
 
   return (
     <Space size={46} direction="vertical">
+      <InputFormCard label="Stake Now" suffix="DEV" onSubmitStake={handleSubmit} />
       <WithdrawCard
         label="Staking"
         onSubmitWithdraw={handleWithdrawStaking}
@@ -33,7 +42,6 @@ export const TransactionForm = ({ propertyAddress }: Props) => {
         amount={myHolderAmount}
         lastUpdate={data?.allocator_allocation_result[0].block_number}
       />
-      <InputFormCard label="Stake Now" suffix="DEV" onSubmitStake={() => console.log('submit')} />
     </Space>
   )
 }
